project/requirements.txt
# Webフレームワーク
Flask==3.0.0
Werkzeug==3.0.1

# データベース
psycopg2-binary==2.9.9

# AI / 機械学習
openai==1.12.0
scikit-learn==1.3.2
numpy==1.26.2

# データ処理
pandas==2.1.4

# 環境変数
python-dotenv==1.0.0

# pgvector（PostgreSQL拡張 - 別途インストールが必要）
# pgvectorのPythonバインディング
pgvector==0.2.0

# その他
uuid==1.30
